- server side clustering with streams - saving location after dragging ends - Possibility to hide other markers

+++ src/main/java/com/hj/geoMapping/common/Marker.java
@@ -27,12 +27,12 @@ public class Marker {
 
 
     @JsonProperty
-    public float getLatitude() {
+    public double getLatitude() {
         return location.getLatitude();
     }
 
     @JsonProperty
-    public float getLongitude() {
+    public double getLongitude() {
         return location.getLongitude();
     }
 
@@ -43,7 +43,7 @@ public class Marker {
     }
 
 
-    public float distanceTo(Marker marker) {
+    public double distanceTo(Marker marker) {
         return location.distanceTo(marker.location);
     }
